200422.19:01
Исправил баги системы жизней
Сделал основу для эффектов игрока
Переделал IWalkable и ILives под классы BaseMovement и LivesBase соответственно
Добавил возможность не скипать диалог

+++ Assets/Скрипты/Сохранения/SceneData.cs
@@ -75,13 +75,12 @@ public class SceneData : MonoBehaviour
     public static void Save()
     {
         var active = Active;
-
         var temp = Data.achievements;
 
         Data = new SaveData
         {
             level = GetActiveScene().buildIndex,
-            lives = (int)active.lives.LivesRemaining,
+            lives = active.lives.LivesRemaining,
             achievements = temp
         };
